Feat: Add call-to-action section in catalog page
