+++ src/components/layout/Header.tsx
@@ -1,32 +1,46 @@
 'use client';
 import { useState } from 'react';
 import Image from 'next/image';
+import Link from 'next/link';
 
 const nav = [
     { label: 'Додаткові послуги', href: '#' },
-    { label: 'Каталог', href: '#' },
+    { label: 'Каталог', href: '/catalog' },
     { label: 'Залишити заявку', href: '#' },
 ];
 
 export default function Header() {
     const [open, setOpen] = useState(false);
     return (
-        <header className="header-blur fixed top-0 left-0 w-full z-30 flex items-center justify-between px-4 py-4 bg-[var(--color-bg)]">
-            <div className="font-bold text-white text-xl">Лого</div>
-            {/* Desktop nav */}
+        <header className="header-blur max-w-8xl mx-auto fixed top-0 left-0 w-full z-30 flex items-center justify-between px-4 py-4 bg-[var(--color-bg)]">
+            <Link
+                href="/"
+                className="font-gotham font-medium text-white text-xl"
+            >
+                Лого
+            </Link>
             <nav className="hidden md:flex gap-8">
-                {nav.map((item) => (
-                    <a
-                        key={item.label}
-                        href={item.href}
-                        className="text-white text-lg font-medium hover:underline underline-offset-4 transition"
-                    >
-                        {item.label}
-                    </a>
-                ))}
+                {nav.map((item) =>
+                    item.href === '/catalog' ? (
+                        <Link
+                            key={item.label}
+                            href={item.href}
+                            className="font-gotham font-medium text-white text-lg hover:underline underline-offset-4 transition"
+                        >
+                            {item.label}
+                        </Link>
+                    ) : (
+                        <a
+                            key={item.label}
+                            href={item.href}
+                            className="font-gotham font-medium text-white text-lg hover:underline underline-offset-4 transition"
+                        >
+                            {item.label}
+                        </a>
+                    )
+                )}
             </nav>
             <div className="flex gap-4 items-center text-white text-2xl">
-                {/* Phone icon - only desktop */}
                 <span className="cursor-pointer hidden md:inline">
                     <Image
                         src="/assets/icons/phone.svg"
@@ -35,7 +49,6 @@ export default function Header() {
                         height={24}
                     />
                 </span>
-                {/* Cart icon - always visible */}
                 <span className="cursor-pointer">
                     <Image
                         src="/assets/icons/cart.svg"
@@ -44,7 +57,6 @@ export default function Header() {
                         height={24}
                     />
                 </span>
-                {/* User icon - only desktop */}
                 <span className="cursor-pointer hidden md:inline">
                     <Image
                         src="/assets/icons/user.svg"
@@ -53,7 +65,6 @@ export default function Header() {
                         height={24}
                     />
                 </span>
-                {/* Burger icon - only mobile */}
                 <button
                     className="md:hidden ml-2"
                     aria-label="Відкрити меню"
@@ -68,11 +79,20 @@ export default function Header() {
                     />
                 </button>
             </div>
-            {/* Mobile menu overlay */}
             {open && (
-                <div className="fixed inset-0 z-50 flex flex-col min-h-screen h-full w-full bg-gradient-to-br from-[#750DBE] to-[#4B0A7A]">
+                <div
+                    className="fixed inset-0 z-50 flex flex-col min-h-screen h-full w-full bg-white"
+                    style={{
+                        backgroundImage: "url('/assets/images/bg2.svg')",
+                        backgroundSize: 'cover',
+                        backgroundPosition: 'center',
+                        backgroundRepeat: 'no-repeat',
+                    }}
+                >
                     <div className="flex items-center justify-between px-4 py-4">
-                        <div className="font-bold text-white text-xl">Лого</div>
+                        <div className="font-gotham font-medium text-white text-xl">
+                            Лого
+                        </div>
                         <div className="flex gap-4 items-center">
                             <span>
                                 <Image
@@ -97,21 +117,32 @@ export default function Header() {
                         </div>
                     </div>
                     <nav className="flex flex-col items-center justify-center flex-1 gap-10">
-                        {nav.map((item) => (
-                            <a
-                                key={item.label}
-                                href={item.href}
-                                className="text-white text-2xl font-semibold"
-                                onClick={() => setOpen(false)}
-                            >
-                                {item.label}
-                            </a>
-                        ))}
+                        {nav.map((item) =>
+                            item.href === '/catalog' ? (
+                                <Link
+                                    key={item.label}
+                                    href={item.href}
+                                    className="font-gotham font-medium text-white text-2xl"
+                                    onClick={() => setOpen(false)}
+                                >
+                                    {item.label}
+                                </Link>
+                            ) : (
+                                <a
+                                    key={item.label}
+                                    href={item.href}
+                                    className="font-gotham font-medium text-white text-2xl"
+                                    onClick={() => setOpen(false)}
+                                >
+                                    {item.label}
+                                </a>
+                            )
+                        )}
                     </nav>
                     <div className="flex justify-center pb-8">
                         <a
                             href="tel:+380508886010"
-                            className="flex items-center gap-2 bg-white text-black rounded-full px-4 py-2 text-lg font-semibold shadow-md"
+                            className="flex items-center gap-2 bg-white text-black rounded-full px-4 py-2 text-lg font-medium shadow-md"
                         >
                             <Image
                                 src="/assets/icons/phone.svg"
